Some bugfixes. PlaneLost icon. In-/out-bound. Shuffled fields
When a plane is selected the associated thumb is scrolled into
the viewport when not visible.

When a plane is lost, the icon is grayed out.

In-/out-bound detection based on distance. When two successive
distance measurement indicate the plane is approaching, an
inbound icon is shown. Else an outbound icon is displayed.

The fields in the thumb are reorganized in a more logical
grouping.