Database alarms, statistics toggle, performance optimization
Alarms are now stored in the mysql database. This to allow easy
modification of the alarms later through the web interface.

The statistics pane can now be hidden/shown by pressing on the
statistics button in the lower left corner.

The performance of the page has been increased, mainly by
disabling transitions which require a lot of CPU/GPU power.
With 100 planes, on my machine, the load is about 8%
(based on Google Chrome Performance).